Add and use import_routes
