+++ plugins/ShinyCMS/config/routes/admin_route_concerns.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+# ShinyCMS ~ https://shinycms.org
+#
+# Copyright 2009-2021 Denny de la Haye ~ https://denny.me
+#
+# ShinyCMS is free software; you can redistribute it and/or modify it under the terms of the GPL (version 2 or later)
+
+# Route concerns for ShinyCMS admin area
+
+concern :with_paging do
+  get '(page/:page)', action: :index, on: :collection, as: ''
+end
+
+concern :with_search do
+  get :search, action: :search, on: :collection
+end
